TakePicture and Display (#41)
* Take Photo Functionality

Needs to utilize the image

* Add Missing Listener

* Requests Permissions for Android 7+

Camera Request Feature is required on most recent APIs. Additionally checks to ensure a camera app exists.

* Minor Dependenacy Update

* More work needed

* Removed Username Temporarily

* Profile Updates

Adds some profile updates

* Profile Updates - Firebase Connection

* working

* Profile Updates

* Profile Updates

* BugFix: Logical Profile Issues

* Forces Navbar to update every load

* Display Image preview

* TakePicture and Display

Adds functionality to take a picture and display it.

* TextOCR Scannable

* OCR Scanner & Username Remembered

OCR Scanner for image. Also adds the email to a share preferances so it can be remembered.

* Camera Rotation + Phone Authentication FrontEnd

Phone Auth Front

* patch v24 issue

* Line Item Bounding

Uses the Textblock bounds and its components Lines for linking items to there names.

* Improved Accuracy

Now need to check for duplicate scans then parse

* Raw Items Parsed

Now need to add these as items to a visual screen for corrections

* Finding Items

* Comments

* Levenshtein Algorithm

Adds an non called algorithm to check distance between a target word and an arguement.

* Layout Skeleton

* more layout